tests/mobile/test_mobile_app.py
"""
Mobile application tests
Note: Requires Appium server and mobile device/emulator
"""
import pytest
from appium import webdriver
from appium.options.common import AppiumOptions
from framework.utils.config_loader import ConfigLoader
from loguru import logger


@pytest.mark.mobile
@pytest.mark.slow
class TestMobileApp:
    """Test cases for mobile application"""
    
    @pytest.fixture(scope="class")
    def mobile_driver(self):
        """Setup mobile driver"""
        config = ConfigLoader().load_config()
        mobile_config = config['mobile']
        appium_url = mobile_config['appium']['server_url']
        
        # iOS configuration
        ios_options = AppiumOptions()
        ios_options.platform_name = "iOS"
        ios_options.platform_version = mobile_config['ios']['platform_version']
        ios_options.device_name = mobile_config['ios']['device_name']
        ios_options.app = mobile_config['ios']['app_path']
        ios_options.automation_name = "XCUITest"
        
        # Android configuration
        android_options = AppiumOptions()
        android_options.platform_name = "Android"
        android_options.platform_version = mobile_config['android']['platform_version']
        android_options.device_name = mobile_config['android']['device_name']
        android_options.app = mobile_config['android']['app_path']
        android_options.automation_name = "UiAutomator2"
        
        # Select platform (default to iOS for example)
        options = ios_options
        
        try:
            driver = webdriver.Remote(appium_url, options=options)
            logger.info("Mobile driver initialized")
            yield driver
            driver.quit()
        except Exception as e:
            logger.error(f"Failed to initialize mobile driver: {e}")
            pytest.skip(f"Mobile testing not available: {e}")
    
    def test_app_launches(self, mobile_driver):
        """Test that app launches successfully"""
        assert mobile_driver is not None, "Mobile driver should be initialized"
        logger.info("Mobile app launched successfully")
    
    def test_caption_interface_displayed(self, mobile_driver):
        """Test that caption interface is displayed"""
        # This would test the actual caption interface
        # Adjust selectors based on actual app structure
        try:
            # Example: Find caption display element
            # caption_element = mobile_driver.find_element("id", "caption_display")
            # assert caption_element.is_displayed()
            logger.info("Caption interface test - adjust based on actual app")
        except Exception as e:
            logger.warning(f"Caption interface test skipped: {e}")
    
    def test_call_functionality(self, mobile_driver):
        """Test call functionality in mobile app"""
        # This would test making a call through the app
        logger.info("Call functionality test - implement based on actual app")
